Moving non-Orchard ASP.NET Core features to a new project
