+++ Lombiq.HelpfulLibraries.AspNetCore/Docs/Localization.md
@@ -1,0 +1,6 @@
+# Lombiq Helpful Libraries - ASP.NET Core Libraries - Localization for Orchard Core
+
+
+
+- `HttpContext` extensions: Adds extension methods to `HttpContext` objects that can be used to fetch localization-specific values such as the current request's UI culture.
+- `LocalizedHtmlString` extensions: Adds extension methods to `LocalizedHtmlString` objects to make view localization more convenient.
