+++ Lombiq.HelpfulLibraries.AspNetCore/Docs/Mvc.md
@@ -1,0 +1,11 @@
+# Lombiq Helpful Libraries - ASP.NET Core Libraries - MVC for Orchard Core
+
+
+
+- `ActionResultHelpers`: For returning common specialized `IActionResult`s from `Controller`s, such as `ZipFile`.
+- Attributes for controllers:
+    - `DevelopmentOnlyAttribute`: Enforces the Development environment on controllers.
+    - `DevelopmentAndLocalhostOnlyAttribute`: Enforces the Development environment as well as localhost.
+- `Controller` extensions: Adds extension methods to `Controller` objects like `.RedirectToLocal(redirectUrl)`.
+- `ResultExecutingContextExtensions` to get some shortcuts to usual context operations in `IAsyncResultFilter`s.
+-
