+++ Lombiq.HelpfulLibraries.AspNetCore/Extensions/EnvironmentHttpContextExtensions.cs
