+++ Lombiq.HelpfulLibraries.AspNetCore/Extensions/TemporaryResponseWrapper.cs
