+++ Lombiq.HelpfulLibraries.AspNetCore/Localization/LocalizationHttpContextExtensions.cs
