+++ Lombiq.HelpfulLibraries.AspNetCore/Middlewares/DeferredTaskMiddleware.cs
@@ -3,7 +3,7 @@ using Microsoft.AspNetCore.Http;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace Lombiq.HelpfulLibraries.Libraries.Middlewares;
+namespace Lombiq.HelpfulLibraries.AspNetCore.Middlewares;
 
 public class DeferredTaskMiddleware
 {
