+++ Lombiq.HelpfulLibraries.AspNetCore/Mvc/DevelopmentOnlyAttribute.cs
@@ -4,19 +4,18 @@ using Microsoft.AspNetCore.Mvc.Filters;
 using Microsoft.Extensions.Hosting;
 using System;
 
-namespace Lombiq.HelpfulLibraries.Libraries.Mvc;
+namespace Lombiq.HelpfulLibraries.AspNetCore.Mvc;
 
 /// <summary>
-/// Enforces the Development environment as well as localhost. When put on a controller or an action it'll set a <see
-/// cref="NotFoundResult"/> if the current <see cref="IHostEnvironment"/> is not Development or if the host of the
-/// current URL is "localhost".
+/// Enforces the Development environment. When put on a controller or an action it'll set a <see cref="NotFoundResult"/>
+/// if the current <see cref="IHostEnvironment"/> is not Development.
 /// </summary>
 [AttributeUsage(AttributeTargets.Class | AttributeTargets.Method, AllowMultiple = false, Inherited = true)]
-public sealed class DevelopmentAndLocalhostOnlyAttribute : ActionFilterAttribute
+public sealed class DevelopmentOnlyAttribute : ActionFilterAttribute
 {
     public override void OnActionExecuting(ActionExecutingContext context)
     {
-        if (!context.HttpContext.IsDevelopmentAndLocalhost())
+        if (!context.HttpContext.IsDevelopment())
         {
             context.Result = new NotFoundResult();
             return;
