+++ Lombiq.HelpfulLibraries.AspNetCore/Readme.md
@@ -1,0 +1,19 @@
+# Lombiq Helpful Libraries - ASP.NET Core Libraries
+
+
+
+## About
+
+Some useful extensions and other helpers for ASP.NET Core.
+
+For general details about and on using the Helpful Libraries see the [root Readme](../Readme.md).
+
+
+## Documentation
+
+Please see the inline documentation of each piece of the API to learn more.
+
+- [Extensions](Docs/Extensions.md)
+- [Localization](Docs/Localization.md)
+- [Middlewares](Docs/Middlewares.md)
+- [MVC](Docs/Mvc.md)
