--- Lombiq.HelpfulLibraries/Docs/LocalizationLibraries.md
@@ -1,13 +1,0 @@
-# Dependency Injection Libraries Documentation
-
-
-
-## `HttpContext` extensions
-
-Adds extension methods to `HttpContext` objects that can be used to fetch localization-specific values such as the current request's UI culture.
-
-## `LocalizedHtmlStringExtensions` extensions
-
-Adds extension methods to `LocalizedHtmlString` objects to make view localization more convenient. 
-
-Please see the inline documentation of each piece of the API to learn more.
