+++ Lombiq.HelpfulLibraries/Docs/MvcLibraries.md
@@ -1,33 +1,13 @@
 # MVC Libraries Documentation
 
 
-Please see the inline documentation of each piece of the API to learn more.
 
-## `Controller` extensions
+## Extensions and helpers
 
-Adds extension methods to `Controller` objects like `.RedirectToLocal(redirectUrl)`.
-
-
-## Attributes for controllers
-
-- `DevelopmentOnlyAttribute`: Enforces the Development environment on controllers.
-- `DevelopmentAndLocalhostOnlyAttribute`: Enforces the Development environment as well as localhost.
-
-
-## Controller Helpers
-
-- `ActionResultHelpers`: For returning common specialized `IActionResult`s, such as `ZipFile`.
-
-
-## Filter extensions
- 
-Use `ResultExecutingContextExtensions` to get some shortcuts to usual context operations in `IAsyncResultFilter`s.
-
-
-## Other extensions
-
-- `UrlHelperExtensions`: Adds extension methods to the `@Url` helper, such as `EditContentItemWithTab`.
+-  `Controller` extensions: Adds extension methods to `Controller` objects like `.RedirectToContentDisplay(content)`.
+- `ResultExecutingContextExtensions` to get some shortcuts to usual context operations in `IAsyncResultFilter`s.
 - `ShapeResultExtensions`: Adds extensions methods generating placement strings on your shape description, such as `UseTab`.
+- `UrlHelperExtensions`: Adds extension methods to the `@Url` helper, such as `EditContentItemWithTab`.
 
 
 ## `TypedRoute`
