+++ Lombiq.HelpfulLibraries/Libraries/Mvc/OrchardResultExecutingContextExtensions.cs
@@ -1,9 +1,8 @@
-using Microsoft.AspNetCore.Mvc.RazorPages;
 using OrchardCore.Admin;
 
 namespace Microsoft.AspNetCore.Mvc.Filters;
 
-public static class ResultExecutingContextExtensions
+public static class OrchardResultExecutingContextExtensions
 {
     /// <summary>
     /// Indicates if the current result is a full view rendering result (i.e. where you can properly inject shapes into
@@ -24,14 +23,4 @@ public static class ResultExecutingContextExtensions
     /// </summary>
     public static bool IsAdmin(this ResultExecutingContext context) =>
         AdminAttribute.IsApplied(context.HttpContext);
-
-    /// <summary>
-    /// Indicates if the current result is a full view rendering result (i.e. where you can properly inject shapes into
-    /// the Layout).
-    /// </summary>
-    /// <remarks>
-    /// <para>The URL /Admin/Media/MediaApplication from OrchardCore.Media will be a full view rendering though.</para>
-    /// </remarks>
-    public static bool IsNotFullViewRendering(this ResultExecutingContext context) =>
-        context.Result is not ViewResult and not PageResult;
 }
