+++ Lombiq.HelpfulLibraries/Libraries/Utilities/ContentHttpContextExtensions.cs
@@ -9,26 +9,9 @@ using System.Linq.Expressions;
 
 namespace Microsoft.AspNetCore.Http;
 
-public static class HttpContextExtensions
+public static class ContentHttpContextExtensions
 {
     private const string ContentSessionDataInfix = "_SessionData_";
-
-    /// <summary>
-    /// Sets the cookie with the given name with a maximal expiration time.
-    /// </summary>
-    public static void SetCookieForever(this HttpContext httpContext, string name, string value) =>
-        httpContext.Response.Cookies.Append(name, value, new CookieOptions
-        {
-            Expires = DateTimeOffset.MaxValue,
-            Secure = true,
-            HttpOnly = true,
-        });
-
-    /// <summary>
-    /// Sets the cookie with the given name with a maximal expiration time.
-    /// </summary>
-    public static void SetCookieForever(this IHttpContextAccessor httpContextAccessor, string name, string value) =>
-        httpContextAccessor.HttpContext.SetCookieForever(name, value);
 
     public static void SetContentSessionData(this HttpContext httpContext, IContent content, object data) =>
         httpContext.Items[GetContentSessionDataKey(content)] = data;
