+++ Readme.md
@@ -1,4 +1,4 @@
-# Lombiq Helpful Libraries for Orchard Core
+# Lombiq Helpful Libraries
 
 
 
@@ -10,7 +10,7 @@
 
 ## About
 
-Various libraries that can be handy when developing for Orchard Core CMS, to be used from your own Orchard modules.
+Various useful libraries that can be handy when developing for .NET, ASP.NET Core, and [Orchard Core CMS](https://orchardcore.net/), to be used from your own projects.
 
 Do you want to quickly try out this project and see it in action? Check it out in our [Open-Source Orchard Core Extensions](https://github.com/Lombiq/Open-Source-Orchard-Core-Extensions) full Orchard Core solution and also see our other useful Orchard Core-related open-source projects!
 
@@ -19,18 +19,18 @@ Note that this project has an Orchard 1 version in the [dev-orchard-1 branch](ht
 
 ## Documentation
 
-The module consists of the following independent libraries (all in their own features):
+Helpful Libraries consists of the following independent libraries (all in their own projects/packages):
 
-- [ASP.NET Core Libraries](Lombiq.HelpfulLibraries/Docs/AspNetCoreLibraries.md)
+- [ASP.NET Core Libraries](Lombiq.HelpfulLibraries.AspNetCore/Readme.md)
+
+To be factored out:
 - [Contents Libraries](Lombiq.HelpfulLibraries/Docs/ContentsLibraries.md)
 - [Database Libraries](Lombiq.HelpfulLibraries/Docs/DatabaseLibraries.md)
 - [DateTime Libraries](Lombiq.HelpfulLibraries/Docs/DateTimeLibraries.md)
 - [Dependency Injection Libraries](Lombiq.HelpfulLibraries/Docs/DependencyInjectionLibraries.md)
-- [Localization Libraries](Lombiq.HelpfulLibraries/Docs/LocalizationLibraries.md)
 - [Method Provider Libraries](Lombiq.HelpfulLibraries/Docs/MethodProviderLibraries.md)
 - [Shapes Libraries](Lombiq.HelpfulLibraries/Docs/ShapesLibraries.md)
 - [Validation Libraries](Lombiq.HelpfulLibraries/Docs/ValidationLibraries.md)
-- [Middlewares](Lombiq.HelpfulLibraries/Docs/Middlewares.md)
 - [MVC Libraries](Lombiq.HelpfulLibraries/Docs/MvcLibraries.md)
 - [Navigation](Lombiq.HelpfulLibraries/Docs/Navigation.md)
 - [Resource Management Libraries](Lombiq.HelpfulLibraries/Docs/ResourceManagementLibraries.md)
@@ -45,7 +45,7 @@ Public APIs are always documented so please always read method comments.
 
 Check out the [`Lombiq.HelpfulLibraries.Samples` project](Lombiq.HelpfulLibraries.Samples) for various examples.
 
-Note: The `Lombiq.HelpfulLibraries.sln` exists to provide centralized building of the independent library projects.
+Note: The `Lombiq.HelpfulLibraries.sln` file exists to provide centralized building of the independent library projects but otherwise you don't need to use it.
 
 ## Contributing and support
